# project-irequest-service/swagger.json
{"openapi":"3.0.0","components":{"examples":{},"headers":{},"parameters":{},"requestBodies":{},"responses":{},"schemas":{"MasterValueHeaderResponse":{"properties":{"id":{"type":"number","format":"double"},"uuid":{"type":"string"},"masterValueCode":{"type":"string"},"masterValueName":{"type":"string"}},"type":"object","additionalProperties":false},"SaveMasterValueHeaderRequest":{"properties":{"id":{"type":"number","format":"double"},"masterValueCode":{"type":"string"},"masterValueName":{"type":"string"}},"type":"object","additionalProperties":false},"MasterValueDetailResponse":{"properties":{"id":{"type":"number","format":"double"},"uuid":{"type":"string"},"masterValueDetailCode":{"type":"string"},"descriptionTh":{"type":"string"},"descriptionEn":{"type":"string"},"sequence":{"type":"number","format":"double"},"masterValueHeaderId":{"type":"number","format":"double"}},"type":"object","additionalProperties":false},"SaveMasterValueDetailRequest":{"properties":{"id":{"type":"number","format":"double"},"masterValueDetailCode":{"type":"string"},"descriptionTh":{"type":"string"},"descriptionEn":{"type":"string"},"sequence":{"type":"number","format":"double"},"masterValueHeaderId":{"type":"number","format":"double"}},"type":"object","additionalProperties":false},"GeneralRequestRequest":{"properties":{"studentCode":{"type":"string"},"topic":{"type":"string"},"cause":{"type":"string"}},"type":"object","additionalProperties":false}},"securitySchemes":{}},"info":{"title":"IRequest Api Service","version":"1.0.0","contact":{}},"paths":{"/iram06/save-master-value-header-post":{"post":{"operationId":"SaveMasterValueHeaderPost","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MasterValueHeaderResponse"}}}}},"description":"บันทึกข้อมูล Master Value Header","tags":["IRAM06 - Master Value"],"security":[],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveMasterValueHeaderRequest"}}}}}},"/iram06/save-master-value-detail-post":{"post":{"operationId":"SaveMasterValueDetailPost","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MasterValueDetailResponse"}}}}},"description":"บันทึกข้อมูล Master Value Detail","tags":["IRAM06 - Master Value"],"security":[],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveMasterValueDetailRequest"}}}}}},"/iram06/find-master-value-header-get":{"get":{"operationId":"FindMasterValueHeaderGet","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/MasterValueHeaderResponse"}],"nullable":true}}}}},"description":"ค้นหาข้อมูล Master Value Header ตาม code","tags":["IRAM06 - Master Value"],"security":[],"parameters":[{"in":"query","name":"master-value-header-code","required":true,"schema":{"type":"string"}}]}},"/iram06/find-master-value-detail-get":{"get":{"operationId":"FindMasterValueDetailGet","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/MasterValueDetailResponse"}],"nullable":true}}}}},"description":"ค้นหาข้อมูล Master Value Detail ตาม code","tags":["IRAM06 - Master Value"],"security":[],"parameters":[{"in":"query","name":"master-value-detail-code","required":true,"schema":{"type":"string"}}]}},"/iram06/find-master-value-details-by-header-get":{"get":{"operationId":"FindMasterValueDetailsByHeaderGet","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MasterValueDetailResponse"},"type":"array","nullable":true}}}}},"description":"ค้นหาข้อมูล Master Value Detail ตาม header code เรียงตาม sequence","tags":["IRAM06 - Master Value"],"security":[],"parameters":[{"in":"query","name":"header-code","required":true,"schema":{"type":"string"}}]}},"/irst05/create-request-post":{"post":{"operationId":"CreateRequestPost","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{}}}}},"tags":["IRST05 - General Request"],"security":[],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralRequestRequest"}}}}}}},"servers":[{"url":"/api"}],"default":{"openapi":"3.0.0","components":{"examples":{},"headers":{},"parameters":{},"requestBodies":{},"responses":{},"schemas":{"MasterValueHeaderResponse":{"properties":{"id":{"type":"number","format":"double"},"uuid":{"type":"string"},"masterValueCode":{"type":"string"},"masterValueName":{"type":"string"}},"type":"object","additionalProperties":false},"SaveMasterValueHeaderRequest":{"properties":{"id":{"type":"number","format":"double"},"masterValueCode":{"type":"string"},"masterValueName":{"type":"string"}},"type":"object","additionalProperties":false},"MasterValueDetailResponse":{"properties":{"id":{"type":"number","format":"double"},"uuid":{"type":"string"},"masterValueDetailCode":{"type":"string"},"descriptionTh":{"type":"string"},"descriptionEn":{"type":"string"},"sequence":{"type":"number","format":"double"},"masterValueHeaderId":{"type":"number","format":"double"}},"type":"object","additionalProperties":false},"SaveMasterValueDetailRequest":{"properties":{"id":{"type":"number","format":"double"},"masterValueDetailCode":{"type":"string"},"descriptionTh":{"type":"string"},"descriptionEn":{"type":"string"},"sequence":{"type":"number","format":"double"},"masterValueHeaderId":{"type":"number","format":"double"}},"type":"object","additionalProperties":false},"GeneralRequestRequest":{"properties":{"studentCode":{"type":"string"},"topic":{"type":"string"},"cause":{"type":"string"}},"type":"object","additionalProperties":false}},"securitySchemes":{}},"info":{"title":"IRequest Api Service","version":"1.0.0","contact":{}},"paths":{"/iram06/save-master-value-header-post":{"post":{"operationId":"SaveMasterValueHeaderPost","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MasterValueHeaderResponse"}}}}},"description":"บันทึกข้อมูล Master Value Header","tags":["IRAM06 - Master Value"],"security":[],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveMasterValueHeaderRequest"}}}}}},"/iram06/save-master-value-detail-post":{"post":{"operationId":"SaveMasterValueDetailPost","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MasterValueDetailResponse"}}}}},"description":"บันทึกข้อมูล Master Value Detail","tags":["IRAM06 - Master Value"],"security":[],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveMasterValueDetailRequest"}}}}}},"/iram06/find-master-value-header-get":{"get":{"operationId":"FindMasterValueHeaderGet","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/MasterValueHeaderResponse"}],"nullable":true}}}}},"description":"ค้นหาข้อมูล Master Value Header ตาม code","tags":["IRAM06 - Master Value"],"security":[],"parameters":[{"in":"query","name":"master-value-header-code","required":true,"schema":{"type":"string"}}]}},"/iram06/find-master-value-detail-get":{"get":{"operationId":"FindMasterValueDetailGet","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/MasterValueDetailResponse"}],"nullable":true}}}}},"description":"ค้นหาข้อมูล Master Value Detail ตาม code","tags":["IRAM06 - Master Value"],"security":[],"parameters":[{"in":"query","name":"master-value-detail-code","required":true,"schema":{"type":"string"}}]}},"/iram06/find-master-value-details-by-header-get":{"get":{"operationId":"FindMasterValueDetailsByHeaderGet","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MasterValueDetailResponse"},"type":"array","nullable":true}}}}},"description":"ค้นหาข้อมูล Master Value Detail ตาม header code เรียงตาม sequence","tags":["IRAM06 - Master Value"],"security":[],"parameters":[{"in":"query","name":"header-code","required":true,"schema":{"type":"string"}}]}},"/irst05/create-request-post":{"post":{"operationId":"CreateRequestPost","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{}}}}},"tags":["IRST05 - General Request"],"security":[],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralRequestRequest"}}}}}}},"servers":[{"url":"/api"}]}}